requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
prometheus-client>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
eth-brownie>=1.19.0
eth-abi>=4.0.0
eth-utils>=2.1.0
eth-typing>=3.4.0